requirements.txt
noise==1.1.1
numpy==1.26.1
Pillow==10.1.0
requests==2.31.0
unittest2==1.1.0